+++ Services/Interfaces/IMedicationService.cs
@@ -1,0 +1,13 @@
+using Roshta.Models;
+
+namespace Roshta.Services.Interfaces;
+
+public interface IMedicationService
+{
+    Task<IEnumerable<Medication>> GetAllMedicationsAsync();
+    Task<Medication?> GetMedicationByIdAsync(int id);
+    Task AddMedicationAsync(Medication medication);
+    Task UpdateMedicationAsync(Medication medication);
+    Task DeleteMedicationAsync(int id);
+    Task<bool> MedicationExistsAsync(int id);
+}